build: renaming folder to fix naming issue
